Add files via upload
We definitely need to look at the sectioning and comments in general (I think we should even comment the purpose of each model/the moderators underneath the model itself). I created a new section at the bottom that has all of the models in it so that I could more easily compare them.